Working deferred rending, withing geometry then lighting pass. GBuffers need optimization.

+++ Source/ProspectEngine/Libraries/MaterialLibrary_impl.h
@@ -25,7 +25,7 @@ namespace Prospect
          const Color& diffuse,
          const Color& ambient,
          const Color& specular,
-         float specularPower);
+         int specularPower);
 
       Material GetMaterial(int index);
       int GetMaterialCount() const;
